requirements.txt
streamlit
oopenai>=1.0.0